Fix gce_test.go to run on everyone's computer.
For reasons related to cached .a files on my computer, I was actually able to import a main package into other packages for a while. Either the .a files got cleaned or the recent Go compiler fixed the unintended support, but in any case now everyone can correctly run the gce_test.