Terminate gracefully after receiving a SIGTERM signal.
Added a SIGTERM handler to the client with a configurable
termination grace period (default 20s). During this period
the client will refuse new connections. At the end of the
period the process will exit. This allows in-flight
database operations to complete when running as a sidecar
container in Kubernetes and performing rolling updates.