Check for file descriptor limit on startup.
Make the proxy check for file descriptor limit at startup using rlimit,
and set it to a value which can support 4k connections if required.

Also, add a flag where the user can override the value to a different
one. Useful in cases where the proxy is used to connect to more than one
instance.

Failure to set appropriate rlimits only leads to warning messages being
logged, and the proxy will startup as usual.