Address more review comments.
Also, go ahead with trying to set the rlimit even when the current hard
limit is less than the value that we are trying to set. Processes with
sufficient privileges will be able to alter the hard limit as well.

From linux manpage:
The soft limit is the value that the kernel enforces for the corre‐
sponding resource.  The hard limit acts as a ceiling for the soft
limit: an unprivileged process may set only its soft limit to a value
in the range from 0 up to the hard limit, and (irreversibly) lower
its hard limit.  A privileged process (under Linux: one with the
CAP_SYS_RESOURCE capability in the initial user namespace) may make
arbitrary changes to either limit value.