Use gcloud oauth token before trying application default credentials (#198)
This changes the oauth token source to first try to retrieve auth tokens using `gcloud config config-helper` first before trying application default credentials. This behavior will only exist when the proxy is invoked as a binary, but not when it is used as a library. The intent here is that its likely reasonable for individuals to run the proxy, but if they are using the proxy as a library, we do still want to encourage them to use a service account, since that is more reliable for production configurations