Print out any errors during SIGTERM-caused shutdown
Before this change any error encountered during a forced shutdown via proxyClient.Shutdown will not be printed out, yet when there is an error the Proxy will still exit with a nonzero status. In addition to exiting with a nonzero status, this CL causes the errors to actually be printed out.

Note that log.Fatal(f) exits the program with exit code 1, which is a change from the prior implementation which exited with code 2. Let me know if exit code 2 was particularly important; I will happily log a message and explicitly exit(2).

I noticed this upon reading issue https://github.com/GoogleCloudPlatform/cloudsql-proxy/issues/388.