fix: ensure certificate is valid on refresh
When clients use IAM authorization, an OAuth2 token is inserted into the
local certificate. In a previous commit, that token could expire within
the configured buffer (i.e., 5 minutes), causing the proxy to fail to
refresh its configuration and to report that the "new ephemeral
certificate expires too soon." Because the code always checked the token
expiration and the certificate expiration, this bug would affect both
clients using IAM auth and clients *not* using it.

This commit ensures the token's expiration is only considered when IAM
auth has been enabled, and when it has been enabled, this commit
likewise ensures the token's expiration is well beyond the buffer.

Fixes #643.