chore: always ask for a fresh token from gcloud
When using gcloud credentials, it's necessary to ensure the token's
expiration is far enough in the future. If we use
oauth2.ReuseTokenSource, the returned token may be valid but may also
expire too soon.

This commit ensures that every time we ask for a gcloud token, we get a
newly refreshed one whose expiration is an hour in the future. Given
that the proxy asks for tokens once every ~hour, this change should not
cause any problems.