fix: use new oauth2 token with cert refresh (#648)
* fix: ensure certificate is valid on refresh

When clients use IAM authorization, an OAuth2 token is inserted into the
local certificate. In a previous commit, that token could expire within
the configured buffer (i.e., 5 minutes), causing the proxy to fail to
refresh its configuration and to report that the "new ephemeral
certificate expires too soon." Because the code always checked the token
expiration and the certificate expiration, this bug would affect both
clients using IAM auth and clients *not* using it.

This commit ensures the token's expiration is only considered when IAM
auth has been enabled, and when it has been enabled, this commit
likewise ensures the token's expiration is well beyond the buffer.

Fixes #643.

* chore: always ask for a fresh token from gcloud

When using gcloud credentials, it's necessary to ensure the token's
expiration is far enough in the future. If we use
oauth2.ReuseTokenSource, the returned token may be valid but may also
expire too soon.

This commit ensures that every time we ask for a gcloud token, we get a
newly refreshed one whose expiration is an hour in the future. Given
that the proxy asks for tokens once every ~hour, this change should not
cause any problems.