fix: ensure certificate expiration is correct
This commit is a follow-on to #648 which only partially addressed the
issue reported in #643. In a previous commit, we ensured the associated
OAuth2 token was refreshed before requesting an ephemeral certificate.
However, the oauth2.TokenSource stored in certs.RemoteCertSource would
nonetheless report an unrefreshed token's expiration. As a result,
attempts to refresh the configuration would fail given the unrefreshed
token would expire too soon.

This commit takes a new approach by updating the leaf certificate's
expiration to be the earlier of the ephemeral certificate's expiration
or the token's expiration. It likewise removes any checks of the token's
expiration outside of certs.RemoteCertSource.

Fixes #643.