chore: don't run the race detector on windows
Prior to this commit, test builds with the race detector fail with:

gcc_libinit_windows.c:58:12: error: implicit declaration of function '_beginthread'

This error is likely on account of the CI environment's gcc version not
being the one Go recommends (i.e., TDM-GCC). See
https://github.com/golang/go/issues/12029.

Since we're already running tests with the race detector in other
builds, it seems reasonable to skip it here and avoid CI dependency
problems.