fix: validate instance connections in liveness probe (#995)
This commit changes the liveness probe to validate all instance
connections. If any single cached connection has an error, the liveness
probe will fail, triggering a restart.

In addition, this commit adjusts the example configuration of the
liveness probe to be forgiving of transient failures. Now, we recommend
triggering the liveness prove every 60 seconds with a total of five
attempts (5 minutes in total) to line up with the certificate refresh
buffer.

Fixes #980.