feat: add support for specifying a quota project (#1044)
This commit adds a CLI flag quota_project such that users may provide a
project against which the SQL Admin API quota is calculated. Typically,
this flag will be used for internal development purposes.

Note: there is an API option WithQuotaProject that does the same thing.
However, that option is not compatible with WithHTTPClient and so we
manually add the quota project instead.