fix: resolve data race on closing client
Previously, when a caller would try to close the client that was
serving, there was a data race between the assignment of a listener, and
the removal of the listener within socketMount. This commit ensures that
the listener is created prior to the socket mount and so makes the data
race impossible.