feat: Downscope OAuth2 token included in ephemeral certificate (#1450)
In the past, when using Auto IAM AuthN, the token embedded into the
ephemeral certificate would have a broad scope. This commit ensures that
in most cases, the embedded token has only a sqlservice.login scope.

When users provide a token to launch the proxy, the proxy will use that
token with whatever scopes are attached.

Fixes #1446.