fix: ensure separate token source with auto-iam-authn (#1637)
When a caller passes the --token flag with --auto-iam-authn, the Proxy
will now require a separate --login-token. This changes ensures a token
with SQL Admin scope does not make its way into the ephemeral
certificate. Instead, the value passed to --login-token will be used and
should have only the sqlservice.login scope.