feat: add support for Auto IP
This commit restores a legacy behavior of the v1 Proxy, this time behind
the flag --auto-ip. When the flag is enabled, the Proxy will first
attempt to use a public IP. If no public IP is present, the Proxy will
second attempt to use a private IP. If neither is present, the Proxy
errors as expected.

This new flag is primarily to support legacy behavior and should not be
used in production otherwise.