fix: disallow auto-iam-authn with gcloud-auth
This combination is insecure because it puts an OAuth2 token with
broader scoped access than necessary into the ephemeral certificate.

Because gcloud-auth is a legacy flag, it is generally discouraged. This
commit also adds a note to make that clear and provides a clear
alternative using Application Default Credentials.

Fixes #1754