chore: `LABEL` Dockerfiles with URL of source repo
Currently, the `cloud-sql-proxy` Dockerfiles that are built and
published on GCR do not have any `LABEL` statements. In my use case of
the `cloud-sql-proxy` image, I want to pin the image to a specific
version and use GitHub Dependabot to periodically update the image to
the latest version. Dependabot currently cannot check for updates to the
`cloud-sql-proxy` image because it does not have the
`org.opencontainers.image.source` label.

This commit adds the line

```dockerfile
LABEL org.opencontainers.image.source="https://github.com/GoogleCloudPlatform/cloud-sql-proxy"
```

to the Dockerfiles that specify the image contents in the
`cloud-sql-proxy` repository. This has the effect of allowing Dependabot
to track updates to this image as documented
[here](https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#docker).