chore: expose additional CLI options in Go interface
The internal users of the Go API prior to this commit had to mix Go
options with CLI string arguments to achieve the desired invocation
(e.g., WithLogger and "--max-connections"). This results in a brittle
API that is error prone.

This commit moves all the internally used options into the programmatic
API such that internal callers can invoke the Proxy without passing any
string based arguments.