fix: correct CI/CD build error
A recent change in the CI/CD infrastructure caused the last release to
miscompile all binaries as Linux x86-64 with dynamic linking. We have
corrected the error in our infrastructure and are cutting a new release
to signal to downstream consumers that the last version is invalid for
all but Linux x86-64.

Fixes #2153