docs: Update kubernetes examples to use K8s native sidecar containers. (#2371)
Running the proxy as a native sidecar container is now the preferred way to run the proxy. Using native
sidecars makes the lifecycle much easier to manage:

- Kubernetes will start the proxy container and wait until it is ready before it starts the application container.
- Kubernetes will keep the proxy container running until the application container exits.
- Kubernetes will stop the proxy container after the application container exits.
- The exit status of the proxy will not impact the pod's overall exit status.

This is especially important for Job and CronJob workloads. Native sidecar containers are supported by 
default in Kuberenetes 1.29. See K8s Sidecar Container:
https://kubernetes.io/docs/concepts/workloads/pods/sidecar-containers/