feat: Add MySQL caching_sha2_password support for the proxy clients using unix sockets. (#2489)
Automatically set the MDX client_protocol_type field to "TCP" or "UDS" for mysql instances that
support Metadata Exchange. This will ensure that password login works correctly in MySQL 8.0 or
higher when the user has a caching_sha2_password.

We will leave this in a draft state until back-end support for this is fully rolled out.

Fixes #2317