feat: Add shutdown command to call the /quitquitquit endpoint. (#2514)
This adds the `shutdown` command the proxy, which will cause a proxy running on the same host to shut down. It 
invokes the shutdown command by sending an HTTP Post to `http://127.0.0.1:9101/quitquitquit`. 

This is especially useful as a preStop hook in Kubernetes. Run  the proxy container with the --quitquitquit and --admin-port flags. Then configure the pre-stop hook with the same admin port. 

```
        # Configure kubernetes to call run the cloud-sql-proxy shutdown command
        # before sending SIGTERM to the proxy when stopping the pod. This will
        # give the proxy more time to gracefully exit.
        lifecycle:
          preStop:
            exec:
              command: ["/cloud-sql-proxy","shutdown", "--admin-port","<ADMIN_PORT>"]
```

Fixes #2511